Refactoring other segmentation models
