+++ Assets/AXIP/AILIA-MODELS/ImageSegmentation/AiliaImageSegmentationSample.cs
@@ -10,22 +10,22 @@ using ailia;
 
 namespace ailiaSDK
 {
+	// Model list
+	public enum ImageSegmentaionModels
+	{
+		HRNetV2_W18_Small_v2,
+		HRNetV2_W18_Small_v1,
+		HRNetV2_W48,
+		hair_segmentation,
+		pspnet_hair_segmentation,
+		deeplabv3,
+		u2net,
+		modnet,
+		segment_anything1
+	}
+
 	public class AiliaImageSegmentationSample : MonoBehaviour
 	{
-		// Model list
-		public enum ImageSegmentaionModels
-		{
-			HRNetV2_W18_Small_v2,
-			HRNetV2_W18_Small_v1,
-			HRNetV2_W48,
-			hair_segmentation,
-			pspnet_hair_segmentation,
-			deeplabv3,
-			u2net,
-			modnet,
-			segment_anything1
-		}
-
 		// Settings
 		public ImageSegmentaionModels imageSegmentaionModels = ImageSegmentaionModels.HRNetV2_W18_Small_v2;
 		public bool gpu_mode = false;
@@ -52,7 +52,6 @@ namespace ailiaSDK
 		int channelFirstUpsideDownKernel;
 
 		// AILIA
-		private AiliaModel ailiaModel;
 		private AiliaCamera ailia_camera = new AiliaCamera();
 
 		// Input source
@@ -74,27 +73,20 @@ namespace ailiaSDK
 		int blendVFlipId;
 
 		// Model input and output
-		int InputWidth;
-		int InputHeight;
-		int InputChannel;
-		int OutputWidth;
-		int OutputHeight;
-		int OutputChannel;
+
 		Texture2D labelTexture;
 		Texture2D originalTexture;
 		Vector2 rawImageSize;
-		float[] output;
-		float[] input;
-		Color32[] outputImage;
-		Color32[] colorPalette = AiliaImageUtil.CreatePalette(256, 127);
-
+	
 		// Segment Anything Model
+		private SegmentationModel segModel;
 		private SegmentAnythingModel samModel;
 		private bool isDraggingForBox = false;
 		private Rect boxRect = new ();
 
 		bool modelPrepared = false;
 		bool modelAllocated = false;
+		string envName = "";
 
 		async void Start()
 		{
@@ -127,38 +119,15 @@ namespace ailiaSDK
 		void AiliaInit()
 		{
 			// Create Ailia
-			ailiaModel = CreateAiliaNet(imageSegmentaionModels, gpu_mode);
+			CreateAiliaNet(imageSegmentaionModels, gpu_mode);
             
 			// Load sample image
 			LoadImage(imageSegmentaionModels, AiliaImageSource);
-		}
 
-		void AllocateInputAndOutputTensor()
-		{
-			float rawImageRatio = rawImageSize.x / rawImageSize.y;
-			float ratio = AiliaImageSource.Width / (float)AiliaImageSource.Height;
-			raw_image.rectTransform.sizeDelta = new Vector2(ratio / rawImageRatio * rawImageSize.x, rawImageSize.y);
-
-			SetShape(imageSegmentaionModels);
-
-			// texture & buffer allocate
-			if (false){//imageSegmentaionModels == ImageSegmentaionModels.segment_anything1){
-				// 元画像サイズ
-				InputWidth = AiliaImageSource.Width;
-				InputHeight = AiliaImageSource.Height;
-				OutputWidth = InputWidth;
-				OutputHeight = InputHeight;
-				raw_image.rectTransform.sizeDelta = new Vector2(1.0f, 1.0f);
-			} else {
-				// 事前にリサイズ
-				AiliaImageSource.Resize(InputWidth, InputHeight);
+			// Resize
+			if (imageSegmentaionModels == ImageSegmentaionModels.segment_anything1){
+				AiliaImageSource.Resize(1024, 1024);
 			}
-
-			input = new float[InputWidth * InputHeight * InputChannel];
-
-			output = new float[OutputWidth * OutputHeight * OutputChannel];
-			outputImage = new Color32[OutputWidth * OutputHeight];
-			labelTexture = new Texture2D(OutputWidth, OutputHeight, TextureFormat.RGBA32, false);
 		}
 
 		void UISetup()
@@ -179,8 +148,17 @@ namespace ailiaSDK
 				mode_text.text = "ailia Image Segmentation";
 			}
 		}
+		Color32 [] VerticalFlip(Color32[] inputImage, int InputWidth, int InputHeight){
+			Color32[] outputImage = new Color32[InputWidth * InputHeight];
+			for(int y=0;y<InputHeight;y++){
+				for(int x=0;x<InputWidth;x++){
+					outputImage[(InputHeight-1-y)*InputHeight+x]=inputImage[y*InputWidth+x];
+				}
+			}
+			return outputImage;
+		}
 
-		async void Update()
+		void Update()
 		{
             HandleClick(Input.GetMouseButton(0), Input.GetMouseButton(1), Input.GetMouseButton(2));
 
@@ -190,7 +168,9 @@ namespace ailiaSDK
 			}
 			if (modelPrepared && !modelAllocated)
 			{
-				AllocateInputAndOutputTensor();
+				if (imageSegmentaionModels != ImageSegmentaionModels.segment_anything1){
+					segModel.AllocateInputAndOutputTensor(imageSegmentaionModels, AiliaImageSource.Width, AiliaImageSource.Height);
+				}
 				modelAllocated = true;
 			}
 			if (camera_mode && !ailia_camera.IsEnable())
@@ -215,7 +195,6 @@ namespace ailiaSDK
 			}
 
 			// Make input data
-			Rect rect = new Rect(0, 0, InputWidth, InputHeight);
 			long start_time = DateTime.UtcNow.Ticks / TimeSpan.TicksPerMillisecond;
 			Color32[] inputImage = null;
 			int inputImageWidth = 0;
@@ -224,14 +203,13 @@ namespace ailiaSDK
 				inputImageWidth = ailia_camera.GetWidth();
 				inputImageHeight = ailia_camera.GetHeight();
 				inputImage = ailia_camera.GetPixels32(); // Bottom2Top format
-				inputImage = ResizeImage(inputImage, inputImageWidth, inputImageHeight, InputWidth, InputHeight);	// Convert to Top2Bottom format
-				inputImageWidth = InputWidth;
-				inputImageHeight = InputHeight;
+				inputImage = VerticalFlip(inputImage, inputImageWidth, inputImageHeight); // Top2Bottom format
 			}else{
-				bool convert_to_top2bottom = true;	// Convert to Top2Bottom format
-				inputImageWidth = InputWidth;
-				inputImageHeight = InputHeight;
-				inputImage = AiliaImageSource.GetPixels32(rect, convert_to_top2bottom);
+				bool convert_to_top2bottom = true;
+				inputImageWidth = AiliaImageSource.Width;
+				inputImageHeight = AiliaImageSource.Height;
+				Rect rect = new Rect(0, 0, inputImageWidth, inputImageHeight);
+				inputImage = AiliaImageSource.GetPixels32(rect, convert_to_top2bottom); // Top2Bottom format
 			}
 
 			long end_time = DateTime.UtcNow.Ticks / TimeSpan.TicksPerMillisecond;
@@ -244,7 +222,7 @@ namespace ailiaSDK
 			{
 				if (samModel.GetClickPoints(0).Length == 0)
 				{
-					samModel.AddClickPoint(InputWidth / 4, InputHeight / 4 + 30);
+					samModel.AddClickPoint(inputImageWidth / 4, inputImageHeight / 4 + 30);
 				}
 
 				samModel.ProcessFrame(inputImage, inputImageWidth, inputImageHeight);
@@ -252,13 +230,7 @@ namespace ailiaSDK
             }
 			else
 			{
-				if(camera_mode){
-					InputDataProcessing(imageSegmentaionModels, inputImage, input);
-				} else {
-					InputDataProcessing(imageSegmentaionModels, inputImage, input);
-				}
-
-				result = ailiaModel.Predict(output, input);
+				result = segModel.ProcessFrame(imageSegmentaionModels, inputImage, inputImageWidth, inputImageHeight);
             }
 
 
@@ -269,31 +241,33 @@ namespace ailiaSDK
 			}
 
 			// convert result to image
-			if (imageSegmentaionModels == ImageSegmentaionModels.hair_segmentation)
-				LabelPaintHairSegmentation(output, outputImage, Color.red);
-			else if (imageSegmentaionModels == ImageSegmentaionModels.pspnet_hair_segmentation)
-				LabelPaintPspnet(output, outputImage, Color.red);
-			else if (imageSegmentaionModels == ImageSegmentaionModels.u2net || imageSegmentaionModels == ImageSegmentaionModels.modnet)
-				LabelPaintU2netModnet(output, outputImage, Color.green);
-			else if (imageSegmentaionModels == ImageSegmentaionModels.segment_anything1)
+			Color32[] outputImage;
+			int outputWidth, outputHeight;
+			if (imageSegmentaionModels == ImageSegmentaionModels.segment_anything1)
 			{
 				outputImage = samModel.visualizedResult.GetPixels32();
-				Debug.Log("InputWidth" + InputWidth + "/" + InputHeight);
+				outputWidth = samModel.visualizedResult.width;
+				outputHeight = samModel.visualizedResult.height;
+				Debug.Log("InputWidth" + inputImageWidth + "/" + inputImageHeight);
 				Debug.Log("samModel.visualizedResult" + samModel.visualizedResult.width + "/" + samModel.visualizedResult.height);
 			}
-			else
-				LabelPaint(output, outputImage, OutputChannel, colorPalette);
+			else{
+				(outputImage, outputWidth, outputHeight) = segModel.PostProcesss(imageSegmentaionModels, inputImageWidth, inputImageHeight);
+			}
 
 			long end_time2 = DateTime.UtcNow.Ticks / TimeSpan.TicksPerMillisecond;
 
 			if (label_text != null)
 			{
-				label_text.text = ((end_time - start_time) + (end_time2 - start_time2)).ToString() + "ms\n" + ailiaModel.EnvironmentName();
+				label_text.text = ((end_time - start_time) + (end_time2 - start_time2)).ToString() + "ms\n" + envName;
 			}
 
 			// for viewer
 			if(originalTexture == null){
-				originalTexture = new Texture2D(InputWidth, InputHeight, TextureFormat.RGBA32, false);
+				originalTexture = new Texture2D(outputWidth, outputHeight, TextureFormat.RGBA32, false);
+			}
+			if(labelTexture == null){
+				labelTexture = new Texture2D(outputWidth, outputHeight, TextureFormat.RGBA32, false);
 			}
 			originalTexture.SetPixels32(inputImage);
 			originalTexture.Apply();
@@ -310,69 +284,19 @@ namespace ailiaSDK
 			blendMaterial.SetFloat(mainVFlipId, 1);	//originalTexture is Top2Bottom
 			blendMaterial.SetFloat(blendVFlipId, 1);	//outputImage is Top2Bottom
 
+
+			float rawImageRatio = rawImageSize.x / rawImageSize.y;
+			float ratio = inputImageWidth/ (float)inputImageHeight;
+			raw_image.rectTransform.sizeDelta = new Vector2(ratio / rawImageRatio * rawImageSize.x, rawImageSize.y);
+
 			raw_image.gameObject.SetActive(true);
 
 			oneshot = false;
 		}
 
 		// Download models and Create ailiaModel
-		AiliaModel CreateAiliaNet(ImageSegmentaionModels modelType, bool gpu_mode = true)
+		void CreateAiliaNet(ImageSegmentaionModels modelType, bool gpu_mode = true)
 		{
-			string asset_path = Application.temporaryCachePath;
-			string serverFolderName = "";
-			string prototxtName = "";
-			string onnxName = "";
-			switch (modelType)
-			{
-				case ImageSegmentaionModels.HRNetV2_W18_Small_v2:
-					serverFolderName = "hrnet";
-					prototxtName = "HRNetV2-W18-Small-v2.onnx.prototxt";
-					onnxName = "HRNetV2-W18-Small-v2.onnx";
-					break;
-				case ImageSegmentaionModels.HRNetV2_W18_Small_v1:
-					serverFolderName = "hrnet";
-					prototxtName = "HRNetV2-W18-Small-v1.onnx.prototxt";
-					onnxName = "HRNetV2-W18-Small-v1.onnx";
-					break;
-				case ImageSegmentaionModels.HRNetV2_W48:
-					serverFolderName = "hrnet";
-					prototxtName = "HRNetV2-W48.onnx.prototxt";
-					onnxName = "HRNetV2-W48.onnx";
-					break;
-				case ImageSegmentaionModels.hair_segmentation:
-					serverFolderName = "hair_segmentation";
-					prototxtName = "hair_segmentation.opt.onnx.prototxt";
-					onnxName = "hair_segmentation.opt.onnx";
-					break;
-				case ImageSegmentaionModels.pspnet_hair_segmentation:
-					serverFolderName = "pspnet-hair-segmentation";
-					prototxtName = "pspnet-hair-segmentation.onnx.prototxt";
-					onnxName = "pspnet-hair-segmentation.onnx";
-					break;
-				case ImageSegmentaionModels.deeplabv3:
-					serverFolderName = "deeplabv3";
-					prototxtName = "deeplabv3.opt.onnx.prototxt";
-					onnxName = "deeplabv3.opt.onnx";
-					break;
-				case ImageSegmentaionModels.u2net:
-					serverFolderName = "u2net";
-					prototxtName = "u2net_opset11.onnx.prototxt";
-					onnxName = "u2net_opset11.onnx";
-					break;
-				case ImageSegmentaionModels.modnet:
-					serverFolderName = "modnet";
-					prototxtName = "modnet.opt.onnx.prototxt";
-					onnxName = "modnet.opt.onnx";
-					break;
-			}
-
-			ailiaModel = new AiliaModel();
-			if (gpu_mode)
-			{
-				// call before OpenFile
-				ailiaModel.Environment(Ailia.AILIA_ENVIRONMENT_TYPE_GPU);
-			}
-
 			AiliaDownload ailia_download = new AiliaDownload();
 			ailia_download.DownloaderProgressPanel = UICanvas.transform.Find("DownloaderProgressPanel").gameObject;
 			List<ModelDownloadURL> urlList = null;
@@ -384,104 +308,25 @@ namespace ailiaSDK
 			}
 			else
 			{
-                urlList = new List<ModelDownloadURL>();
-                urlList.Add(new ModelDownloadURL() { folder_path = serverFolderName, file_name = prototxtName });
-                urlList.Add(new ModelDownloadURL() { folder_path = serverFolderName, file_name = onnxName });
+				segModel = new SegmentationModel();
+				urlList = segModel.GetModelURLs(imageSegmentaionModels);
             }
 
             StartCoroutine(ailia_download.DownloadWithProgressFromURL(urlList, () =>
             {
                 if (imageSegmentaionModels == ImageSegmentaionModels.segment_anything1)
                 {
-                    samModel.InitializeModels(gpu_mode);
-                    modelPrepared = samModel.modelsInitialized;
+                    modelPrepared = samModel.InitializeModels(gpu_mode);
+					envName = samModel.EnvironmentName();
                 }
                 else
                 {
-                    modelPrepared = ailiaModel.OpenFile(asset_path + "/" + prototxtName, asset_path + "/" + onnxName);
+                    modelPrepared = segModel.InitializeModels(imageSegmentaionModels, gpu_mode);
+					envName = segModel.EnvironmentName();
                 }
             }));
-
-			return ailiaModel;
 		}
 
-		void SetShape(ImageSegmentaionModels imageSegmentaionModels)
-		{
-			Ailia.AILIAShape shape = null;
-			switch (imageSegmentaionModels)
-			{
-				case ImageSegmentaionModels.HRNetV2_W18_Small_v2:
-				case ImageSegmentaionModels.HRNetV2_W18_Small_v1:
-				case ImageSegmentaionModels.HRNetV2_W48:
-				case ImageSegmentaionModels.pspnet_hair_segmentation:
-				case ImageSegmentaionModels.deeplabv3:
-					shape = ailiaModel.GetInputShape();
-					InputWidth = (int)shape.x;
-					InputHeight = (int)shape.y;
-					InputChannel = (int)shape.z;
-					shape = ailiaModel.GetOutputShape();
-					OutputWidth = (int)shape.x;
-					OutputHeight = (int)shape.y;
-					OutputChannel = (int)shape.z;
-					break;
-
-				case ImageSegmentaionModels.hair_segmentation:
-					shape = new Ailia.AILIAShape();
-					shape.x = 3;
-					shape.y = (uint)AiliaImageSource.Width;
-					shape.z = (uint)AiliaImageSource.Height;
-					shape.w = 1;
-					shape.dim = 4;
-					ailiaModel.SetInputShape(shape);
-					InputWidth = AiliaImageSource.Width;
-					InputHeight = AiliaImageSource.Height;
-					InputChannel = 3;
-					OutputWidth = AiliaImageSource.Width;
-					OutputHeight = AiliaImageSource.Height;
-					OutputChannel = 1;
-					break;
-
-				case ImageSegmentaionModels.u2net:
-					shape = ailiaModel.GetInputShape();
-					InputWidth = (int)shape.x;
-					InputHeight = (int)shape.y;
-					InputChannel = (int)shape.z;
-					shape = ailiaModel.GetOutputShape();
-					OutputWidth = (int)shape.x;
-					OutputHeight = (int)shape.y;
-					OutputChannel = (int)shape.z;
-					break;
-
-                case ImageSegmentaionModels.modnet:
-                    shape = new Ailia.AILIAShape();
-                    shape.x = (uint)AiliaImageSource.Width / 32 * 32;
-                    shape.y = (uint)AiliaImageSource.Height / 32 * 32;
-                    shape.z = 3;
-                    shape.w = 1;
-                    shape.dim = 4;
-                    ailiaModel.SetInputShape(shape);
-                    InputWidth = AiliaImageSource.Width / 32 * 32;
-                    InputHeight = AiliaImageSource.Height / 32 * 32;
-                    InputChannel = 3;
-                    shape = ailiaModel.GetOutputShape();
-                    OutputWidth = (int)shape.x;
-                    OutputHeight = (int)shape.y;
-                    OutputChannel = (int)shape.z;
-                    Debug.Log("" + OutputChannel + "/" + OutputWidth + "/" + OutputHeight);
-                    //OutputWidth = AiliaImageSource.Width/32*32;
-                    //OutputHeight = AiliaImageSource.Height/32*32;
-                    //OutputChannel = 1;
-                    break;
-
-                case ImageSegmentaionModels.segment_anything1:
-                    (InputWidth, InputHeight) = samModel.GetInputSize(AiliaImageSource);
-                    InputChannel = 3;
-					OutputWidth = InputWidth;
-                    OutputHeight = InputHeight;
-                    OutputChannel = 1;
-                    break;
-            }
-		}
 
 		void LoadImage(ImageSegmentaionModels imageSegmentaionModels, AiliaImageSource ailiaImageSource)
 		{
@@ -513,131 +358,6 @@ namespace ailiaSDK
             }
 		}
 
-		Color32 [] ResizeImage(Color32[] inputImage, int tex_width, int tex_height, int InputWidth, int InputHeight){
-			Color32[] outputImage = new Color32[InputWidth * InputHeight];
-			for(int y=0;y<InputHeight;y++){
-				for(int x=0;x<InputWidth;x++){
-					int x2 = x*tex_width/InputWidth;
-					int y2 = y*tex_height/InputHeight;
-					outputImage[(InputHeight-1-y)*InputHeight+x]=inputImage[y2*tex_width+x2];
-				}
-			}
-			return outputImage;
-		}
-
-		void InputDataProcessing(ImageSegmentaionModels imageSegmentaionModels, Color32[] inputImage, float[] processedInputBuffer)
-		{
-			float weight = 1f / 255f;
-			float bias = 0;
-			bool rgbRepeats = false;
-			switch (imageSegmentaionModels)
-			{
-				case ImageSegmentaionModels.hair_segmentation:
-					rgbRepeats = true;
-					break;
-				case ImageSegmentaionModels.pspnet_hair_segmentation:
-					InputDataProcessing_PSP(inputImage, processedInputBuffer);
-					return;
-				case ImageSegmentaionModels.deeplabv3:
-					weight = 1f / 127.5f;
-					bias = -1;
-					break;
-				case ImageSegmentaionModels.modnet:
-					weight = 1f / 127.5f;
-					bias = -1;
-					break;
-				default:
-					break;
-			}
-
-			// flatten input data
-			if (rgbRepeats)
-			{
-				for (int i = 0; i < inputImage.Length; i++)
-				{
-					// rgbrgbrgb...
-					processedInputBuffer[i * 3 + 0] = (inputImage[i].r) * weight + bias;
-					processedInputBuffer[i * 3 + 1] = (inputImage[i].g) * weight + bias;
-					processedInputBuffer[i * 3 + 2] = (inputImage[i].b) * weight + bias;
-				}
-			}
-			else
-			{
-				for (int i = 0; i < inputImage.Length; i++)
-				{
-					// rrr...ggg...bbb...
-					processedInputBuffer[i + inputImage.Length * 0] = (inputImage[i].r) * weight + bias;
-					processedInputBuffer[i + inputImage.Length * 1] = (inputImage[i].g) * weight + bias;
-					processedInputBuffer[i + inputImage.Length * 2] = (inputImage[i].b) * weight + bias;
-				}
-			}
-		}
-
-		void InputDataProcessing_PSP(Color32[] inputImage, float[] processedInputBuffer)
-		{
-			for (int i = 0; i < inputImage.Length; i++)
-			{
-				// rrr...ggg...bbb...
-				processedInputBuffer[i + inputImage.Length * 0] = (inputImage[i].r * (1f / 255f) - 0.485f) * (1f / 0.229f);
-				processedInputBuffer[i + inputImage.Length * 1] = (inputImage[i].g * (1f / 255f) - 0.456f) * (1f / 0.224f);
-				processedInputBuffer[i + inputImage.Length * 2] = (inputImage[i].b * (1f / 255f) - 0.406f) * (1f / 0.225f);
-			}
-		}
-
-		void LabelPaint(float[] labelData, Color32[] pixelBuffer, int channel, Color32[] palette)
-		{
-			Debug.Assert(palette.Length >= channel, "wrong palette size");
-			Debug.Assert(labelData.Length == pixelBuffer.Length * channel, "wrong parameter");
-
-			for (int i = 0; i < pixelBuffer.Length; i++)
-			{
-				int maxIndex = 0;
-				float maxValue = labelData[i + 0 * (pixelBuffer.Length)];
-				for (int c = 1; c < channel; c++)
-				{
-					if (maxValue < labelData[i + c * pixelBuffer.Length])
-					{
-						maxValue = labelData[i + c * pixelBuffer.Length];
-						maxIndex = c;
-					}
-				}
-				pixelBuffer[i] = palette[maxIndex];
-			}
-		}
-
-		void LabelPaintHairSegmentation(float[] labelData, Color32[] pixelBuffer, Color32 color)
-		{
-			Debug.Assert(labelData.Length == pixelBuffer.Length, "wrong parameter");
-
-			for (int i = 0; i < pixelBuffer.Length; i++)
-			{
-				pixelBuffer[i] = color;
-				pixelBuffer[i].a = (byte)(Mathf.Clamp01(labelData[i]) * 255);
-			}
-		}
-
-		void LabelPaintPspnet(float[] labelData, Color32[] pixelBuffer, Color32 color)
-		{
-			Debug.Assert(labelData.Length == pixelBuffer.Length, "wrong parameter");
-
-			for (int i = 0; i < pixelBuffer.Length; i++)
-			{
-				pixelBuffer[i] = color;
-
-				float k = Mathf.Exp(labelData[i]);
-				pixelBuffer[i].a = (byte)(k / (1.0f + k) * 255);
-			}
-		}
-
-		void LabelPaintU2netModnet(float[] labelData, Color32[] pixelBuffer, Color32 color)
-		{
-			Debug.Assert(labelData.Length == pixelBuffer.Length, "wrong parameter");	
-			for (int i = 0; i < pixelBuffer.Length; i++)
-			{
-				pixelBuffer[i] = color;
-				pixelBuffer[i].a = (byte)(255 - Mathf.Clamp01(labelData[i]) * 255);
-			}	 
-		}
 
 		void HandleClick(bool leftClick, bool rightClick, bool middleClick)
 		{
@@ -758,9 +478,14 @@ namespace ailiaSDK
 			DestroyAiliaDetector();
 		}
 
-		private void DestroyAiliaDetector()
+		void DestroyAiliaDetector()
 		{
-			ailiaModel?.Close();
+			if (samModel != null){
+				samModel.Destroy();
+			}
+			if (segModel != null){
+				segModel.Destroy();
+			}
 		}
 	}
 }
