+++ Assets/AXIP/AILIA-MODELS/ImageSegmentation/SegmentationModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 94f7480f2028e498596562d908b78686
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
